Very very basic example of how multiple backstacks might work.
The problem: We may need to handle multiple backstacks to meet the requirements of our future navigation. The current implementation of this in the app is horrifically broken
and very, very hard to work with. I wouldn't even call it an implementation as we are keeping a fragment with a child backstack alive for a prolonged duration. This breaks
the lifecycle of that fragment AND other fragments related to it e.g. Live and Discover. This is not good.

In this solution I have attempted to split the child back stack from the actual fragment and into a Stack in a ViewModel. The relevant Fragment is then responsible for putting the
correct fragment on the screen where it is required. The Stack is now managed in a ViewModel (which will survive fragment recreation), moving away from the Fragment backstack for child
fragments. We should be aware of this ticket in the Google issue tracker: https://issuetracker.google.com/issues/80029773 which will resolve the issue we see with child backstacks
not surviving fragment recreation! However, do we we want to tie ourselves to the fragment backstack for all navigation? Maybe, as long as it is testable.

We still want to use the navigation component for top level navigation (and hopefully sub navigation in the future.